refactor: update import paths for 2024 components and remove unused GetDpPage
