+++ src/components/2024/EventSchedule/Workshop.tsx
@@ -3,7 +3,7 @@ import Image from "next/image";
 import { Button } from "@chakra-ui/react";
 import { BsArrowUpRight, BsInfoCircle } from "react-icons/bs";
 import lanyard from "@public/icons/lanyard-white-bg.svg";
-import { ctaLinks } from "@/data";
+import { ctaLinks } from "@/data/2024";
 import {
   LongEventScheduleCard,
   IntroductionEventScheduleCard,
@@ -14,7 +14,7 @@ import {
   engineeringOneSchedules,
   designSchedules,
   engineeringTwoSchedules,
-} from "@/data/schedule";
+} from "@/data/2024/schedule";
 
 const Workshop: React.FC = () => {
   return (
