+++ src/components/2024/date-and-location/index.tsx
@@ -5,7 +5,7 @@ import Image from "next/image";
 import { Button } from "@chakra-ui/react";
 import { BsArrowUpRight } from "react-icons/bs";
 import Link from "next/link";
-import { ctaLinks } from "@/data";
+import { ctaLinks } from "@/data/2024";
 
 const DateAndLocation = () => {
   return (
